requirements.txt
asgiref==3.5.0
Django==3.2.13
django-jazzmin==2.5.0
django-modeltranslation==0.18.2
django-phone-field==1.8.1
django-random-id-generator==0.1.0
gunicorn==20.1.0
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
validate-email==1.3
